refactor: move tui tests into module file
